Move logic to ResultsTable
